Rename from "cleanlab-cli" into "cleanlab-studio"
