+++ .github/workflows/ci.yml
@@ -27,7 +27,7 @@ jobs:
           pip install --upgrade pip
           pip install .
           pip install -r tests/requirements_test.txt
-      - name: Install Cleanlab CLI
+      - name: Install Cleanlab Studio client
         run: pip install -e .
       - name: Run tests
         run: |
@@ -42,5 +42,5 @@ jobs:
           python -m pip install --upgrade pip
           pip install . # install deps
           pip install mypy
-          mypy --install-types --non-interactive cleanlab_cli
-      - run: mypy --strict cleanlab_cli
+          mypy --install-types --non-interactive cleanlab_studio
+      - run: mypy --strict cleanlab_studio
