+++ cleanlab_studio/cleanset/download_helpers.py
@@ -1,8 +1,8 @@
 import jsonstreams
 from typing import Dict
 
-from cleanlab_cli.types import RecordType, DatasetFileExtension
-from cleanlab_cli.util import get_dataset_file_extension, get_dataset_chunks, append_rows
+from cleanlab_studio.types import RecordType, DatasetFileExtension
+from cleanlab_studio.util import get_dataset_file_extension, get_dataset_chunks, append_rows
 
 
 def combine_fields_with_dataset(
