+++ cleanlab_studio/dataset/commands.py
@@ -1,6 +1,6 @@
 import click
-from cleanlab_cli.dataset.schema import schema
-from cleanlab_cli.dataset.upload import upload
+from cleanlab_studio.dataset.schema import schema
+from cleanlab_studio.dataset.upload import upload
 
 
 @click.group(help="upload datasets, generate & validate dataset schema")
