+++ cleanlab_studio/dataset/schema_types.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import Dict, Set, Optional, Union, Any
 
-from cleanlab_cli.types import Modality
+from cleanlab_studio.types import Modality
 
 SchemaMetadataDictType = Dict[str, Optional[str]]
 SchemaFieldsDictType = Dict[str, Dict[str, str]]
