+++ tests/schema_generation/test_infer_boolean_data_type.py
@@ -1,5 +1,5 @@
-from cleanlab_cli.dataset import DataType, FeatureType
-from cleanlab_cli.dataset.schema_helpers import infer_types
+from cleanlab_studio.dataset import DataType, FeatureType
+from cleanlab_studio.dataset.schema_helpers import infer_types
 
 
 class TestInferBooleanDataType:
