+++ tests/schema_generation/test_infer_string_data_type.py
@@ -2,8 +2,8 @@ import random
 import uuid
 from datetime import datetime, timedelta
 
-from cleanlab_cli.dataset import DataType, FeatureType
-from cleanlab_cli.dataset.schema_helpers import infer_types
+from cleanlab_studio.dataset import DataType, FeatureType
+from cleanlab_studio.dataset.schema_helpers import infer_types
 
 
 class TestInferStringDataType:
